make the interval more strongly typed, add javadoc
The schema has been modified such that the aggregation_template.interval field
is now a tuple. The map key in tenant.retentions is also a tuple now as well.
That tuple consists of the metric type and the interval. At some point we may
wind up replacing these tuples with UDTs.